+++ src/main/java/cn/org/shelly/edu/model/req/ProposalInitReq.java
@@ -3,6 +3,8 @@ package cn.org.shelly.edu.model.req;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProposalInitReq {
 
@@ -10,7 +12,7 @@ public class ProposalInitReq {
     private Long teamId;
 
     @NotNull(message = "初始积分不能为空")
-    private Integer initialScore;
+    private BigDecimal initialScore;
 
     @NotNull
     private Long gameId;
